Class restructure and CALL API complete
Accidently added the pojo and jsonqueue packages in test folder instead of the main so moved it to the right place. made the CallAPI class that gets the queuewrapper class passed to it and then post the serizazed version of the class into the glassfish server. Please add dependencies to build gradle.

    compile group: 'com.google.code.gson', name: 'gson', version: '2.7'
    compile 'org.apache.httpcomponents:httpclient:4.5'
    compile group: 'org.apache.commons', name: 'commons-io', version: '1.3.2'

this throws an erorr in the build gradle but the errors in the code regarding using the proper library disappear. need to furthur investigate which dependencies to use.